+++ src/file_tree.mli
@@ -1,23 +1,34 @@
 open! Import
 
-
 module Dir : sig
   type t
 
   val path     : t -> Path.t
   val files    : t -> String_set.t
   val sub_dirs : t -> t String_map.t
+
+  (** Whether this directory is ignored by a [jbuild-ignore] file in
+      one of its ancestor directories. *)
+  val ignored : t -> bool
+
+  val fold
+    :  t
+    -> traverse_ignored_dirs:bool
+    -> init:'a
+    -> f:(t -> 'a -> 'a)
+    -> 'a
 end
 
 type t
 
-val load : Path.t -> t
+val load : ?extra_ignored_subtrees:Path.Set.t -> Path.t -> t
 
-type 'a fold_callback_result =
-  | Cont            of 'a
-  | Dont_recurse_in of String_set.t * 'a
-
-val fold : t -> init:'a -> f:(Dir.t -> 'a -> 'a fold_callback_result) -> 'a
+val fold
+  :  t
+  -> traverse_ignored_dirs:bool
+  -> init:'a
+  -> f:(Dir.t -> 'a -> 'a)
+  -> 'a
 
 val root : t -> Dir.t