+++ src/jbuild_load.mli
@@ -9,7 +9,6 @@ end
 
 type conf =
   { file_tree : File_tree.t
-  ; tree      : Alias.tree
   ; jbuilds   : Jbuilds.t
   ; packages  : Package.t String_map.t
   }