+++ test/blackbox-tests/test-cases/lib-available/run.t
@@ -1,3 +1,3 @@
-  $ $JBUILDER build -j1 @runtest --root . --debug-dependency-path
+  $ $JBUILDER build -j1 @runtest --root . --debug-dependency-path 2>&1 | sed "s/ cmd /  sh /"
             sh alias runtest
             sh alias runtest